# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API",
        "description": "Documentación de la API proyecto CDMI",
        "contact": {
            "email": "tucorreo@ejemplo.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/donaciones": {
            "get": {
                "tags": [
                    "Donaciones"
                ],
                "summary": "Listar donaciones",
                "description": "Obtiene una lista paginada de todas las donaciones registradas",
                "operationId": "09b45d42e458d712cd419db36c17eeb9",
                "responses": {
                    "200": {
                        "description": "Lista de donaciones paginada"
                    }
                }
            },
            "post": {
                "tags": [
                    "Donaciones"
                ],
                "summary": "Registrar una nueva donación",
                "description": "Crea una nueva donación en la base de datos",
                "operationId": "8e3c28397006ef4e0ea926104079d3af",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "monto",
                                    "fecha"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "monto": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-25"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Donación creada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/donaciones/{id}": {
            "get": {
                "tags": [
                    "Donaciones"
                ],
                "summary": "Obtener una donación específica",
                "description": "Devuelve los datos de una donación específica por su ID",
                "operationId": "2897e88df84592ddc3070e280482af4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la donación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la donación"
                    },
                    "404": {
                        "description": "Donación no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Donaciones"
                ],
                "summary": "Actualizar una donación",
                "description": "Actualiza los datos de una donación existente",
                "operationId": "c7761437f984138672147348ee1ab751",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la donación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "monto",
                                    "fecha"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Ana López"
                                    },
                                    "monto": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.5
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-26"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Donación actualizada exitosamente"
                    },
                    "404": {
                        "description": "Donación no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Donaciones"
                ],
                "summary": "Eliminar una donación",
                "description": "Elimina una donación de la base de datos por su ID",
                "operationId": "2189f9eb7b9488414cf9b38a435ed4f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la donación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Donación eliminada exitosamente"
                    },
                    "404": {
                        "description": "Donación no encontrada"
                    }
                }
            }
        },
...
    },
...
}